Split testruns/tests.py
Split tests.py into separate files for models, views, and data.
